MINI PROJECT -Design and Development of Memory Game to Assess  Psychological Factors of a Student for Career Selection
The main aim of the paper is to assess the memory power of a student for selecting the appropriate memory game career for the student using psychological factors. For this, the expert system is designed to assess the psychological factor of a student to build a cognitive model and identify the most suitable career(S) for the student. For this, it is necessary to assess various psychological factors of the student which include the speed, time taken, and number of clicks used.to assess these psychological factors of a student, the system asks the student to play the memory game a number of time. Basing on the score obtained and by the factors of the student the system builds a cognitive model for the student.in this paper the main focus is to design a memory game by assessing the psychological factors of student for the career guidance of a student.